+++ core/incremental/mod.rs
@@ -5,6 +5,7 @@ pub mod dbsp;
 pub mod expr_compiler;
 pub mod filter_operator;
 pub mod input_operator;
+pub mod join_operator;
 pub mod operator;
 pub mod persistence;
 pub mod project_operator;
